requirements.txt
pillow==9.4.0
numpy==1.24.1

ipython>=8.8.0
black>=22.12.0
flake8>=6.0.0